Making h5 first, then going into turbo_seti. Adding new script for making h5 files. Also, removing h5 at the end to save disk space.

+++ L-band_seti_pipeline.py
@@ -79,28 +79,47 @@ def cmd_tool(args=None):
         else:
             star_path = star.split('spliced')[0].rstrip('/')+'/'
 
+
+        #------------------------------------
+        # Make hdf5
+
+        out,err = reset_outs()
+
+        command=['python','fil2h5.py',star_path+star_name,'-o',out_dir+'stars_out/']
+        print ' '.join(command)
+        proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        (out, err) = proc.communicate()
+
+         err_file = out_dir+'logs/'+star_name.replace('.fil','.make_h5.err')
+         out_file = out_dir+'logs/'+star_name.replace('.fil','.make_h5.log')
+
+        with open(out_file), 'a') as f:
+            f.write(out)
+
+        if err or proc.returncode != 0:
+            with open(err_file, 'a') as f:
+                f.write(err)
+
         #------------------------------------
         #Run turbo_seti
 
         out,err = reset_outs()
 
-        command=['python','/home/eenriquez/software/bl-soft/turbo_seti/bin/seti_event.py',star_path+star_name,'-M','2','-s','20','-c','115,395','-o',out_dir+'stars_out/']
+        command=['python','/home/eenriquez/software/bl-soft/turbo_seti/bin/seti_event.py',out_dir+'stars_out/'+star_name.replace('.fil','.h5'),'-M','2','-s','20','-c','115,395']
         print ' '.join(command)
+        proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        (out, err) = proc.communicate()
 
-        err_file = out_dir+'logs/'+star_name.replace('.fil','.seti_event.err')
-        out_file = out_dir+'logs/'+star_name.replace('.fil','.seti_event.log')
+         err_file = out_dir+'logs/'+star_name.replace('.fil','.seti_event.err')
+         out_file = out_dir+'logs/'+star_name.replace('.fil','.seti_event.log')
+#         try_it = os.system(' '.join(command) +' 2> '+err_file+' 1> '+out_file)
 
-        try_it = os.system(' '.join(command) +' 2> '+err_file+' 1> '+out_file)
+        with open(out_file), 'a') as f:
+            f.write(out)
 
-#         proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-#         (out, err) = proc.communicate()
-#
-#         with open(out_dir+'logs/'+star_name.replace('.fil','.seti_event.log'), 'a') as f:
-#             f.write(out)
-#
-#         if err or proc.returncode != 0:
-#             with open(out_dir+'logs/'+star_name.replace('.fil','.seti_event.err'), 'a') as f:
-#                 f.write(err)
+        if err or proc.returncode != 0:
+            with open(err_file, 'a') as f:
+                f.write(err)
 
         #------------------------------------
         #Add failed one to list.
@@ -111,14 +130,17 @@ def cmd_tool(args=None):
 
                 with open(error_list,'a') as file_fail:
                     file_fail.write(star)
+
+                remove_h5 =False
             else:
                 os.remove(err_file)
+                remove_h5=True
 
         #------------------------------------
         #Delete h5 file
 
-        os.remove(out_dir+'stars_out/'+star_name.replace('.fil','.h5'))
-
+        if remove_h5:
+            os.remove(out_dir+'stars_out/'+star_name.replace('.fil','.h5'))
 
 if __name__ == "__main__":
     cmd_tool()
